+++ src/test/java/com/uvasoftware/scanii/ScaniiClientsTest.java
@@ -1,6 +1,5 @@
 package com.uvasoftware.scanii;
 
-import com.uvasoftware.scanii.batch.ScaniiBatchClient;
 import com.uvasoftware.scanii.impl.DefaultScaniiClient;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
@@ -27,26 +26,5 @@ class ScaniiClientsTest {
     ScaniiClient client = ScaniiClients.createDefault(ScaniiTarget.AUTO, "key", "secret", hc);
     Assertions.assertNotNull(client);
     Assertions.assertEquals(hc, ((DefaultScaniiClient) client).getHttpClient());
-  }
-
-  @Test
-  void createBatch() {
-    ScaniiBatchClient client = ScaniiClients.createBatch(ScaniiTarget.AUTO, "key", "secret");
-    Assertions.assertNotNull(client);
-  }
-
-  @Test
-  void createBatch1() {
-    ScaniiBatchClient client = ScaniiClients.createBatch(ScaniiTarget.AUTO, "key", "secret", 10);
-    Assertions.assertNotNull(client);
-  }
-
-  @Test
-  void createBatch2() {
-    CloseableHttpClient hc = HttpClients.createMinimal();
-    ScaniiBatchClient client = ScaniiClients.createBatch(ScaniiTarget.AUTO, "key", "secret", 32, hc);
-    Assertions.assertNotNull(client);
-    Assertions.assertEquals(hc, ((DefaultScaniiClient) client.getClient()).getHttpClient());
-
   }
 }