+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v4.0.0
 * Dropped support for protocol versions older than 2.1
+* Dropped batch client (it wasn't getting used, and it's simple to build your own)
 * Refactored ScaniiTarget to allow for custom targets. ScaniiTarget.latest() should be replaced with ScaniiTarget.AUTO
 * Added method to create a client without passing a specific target
 * Dropped usage of Google Guava